refactor(application): extract doctorApplicationService
